Make Transformer groups in TreeTransformers arre based.
The main reason for doing this is that it gives us a handle for further optimizations, should they become necessary. Namely, we can then for each subaction of preparing and transforming create an array of indices of those transformation that have a non-default implementation of these actions. Transformations with default implementations can then be simply skipped.