Avoiding cyclic references due to too eager asSeenFrom
Compared to Scala 2, dotc computes the member type together with the symbol in a
denotation when doing a findMember. This can be problematic because it forces the memberType earlier
than before. In fact it leads to Stackoverflows when dealing with F-bounded types.

This commit fixes that by computing the info of some denotations lazily.