Don't force NamedType denotations in containsRefinedThis
containsRefinedThis inspects symbols and infos of named types in order
to avoid needless traversals. As i974 shows, this can lead to infinite
recursions. The fix is not to force a NamedType denotation, assume the
worst and traverse the prefix if a NamedType is not yet populated
with a denotation.

Fixes #974 and makes MutableSortedSetFactory in stdlib compile.