Recursive transparent calls inline only in redex position
Suppress recursive inlining of transparent methods unless the call is in
redex position.

Recursive calls to transparent methods under an if-then-else or match are not
inlined unless the if-then-else or match is reduced away. Recursive calls on the
rhs of a method definition, type definition, or lazy value definition are also
not inlined.

In principle, we should also suppress inlining calls in a by-name parameter position,
but this is harder to do, first because we don't have always have the function type when
deciding this (in typedApply of TransparentTyper), and second, because if the outer
call with the by-name parameter is to a transparent method, we might decide after
inlining that call that the by-name parameter is after all in redex position, so
we should inline it recursively. But by that time, we might already have used
the recusive call's type to instantiate type parameters of the enclosing function.
If the recursive call is not inlined beforehand, these instantiations might lose
precision. Therefore, it's safer to inline transparent calls in by-name parameters anyway.